[REEF-695] Reorganize reef-examples/scheduler classes into packages
Move classes into client and driver packages.
Task classes are in a separate package, o.a.r.examples.library.

JIRA:
  [REEF-695](https://issues.apache.org/jira/browse/REEF-695)

Pull Request:
  Closes #
